Integrating Character Selection and help Menus
This brings the Character Selection screen online. HOWEVER the CharacterSelection and GameManager Need wiring together. There are commented out lines in the CharacterSelection Screen that need to be adjusted.

Win and loss scenes are found in the res://Game/UI/PopUps folder. They just need to be wired / replace the previous ones if needed.

Hero Assets have the sprite sheets (.png) used in the Character selection and the .tres for those are found in the Game/Resources folder/